+++ lib/src/router_context.dart
@@ -4,5 +4,5 @@ class RouterContext {
   RouterContext({this.path, this.params, this.router});
   final Map<String, String> params;
   final String path;
-  final RouterNG router;
+  final Router router;
 }