+++ test/codecov.dart
@@ -3,8 +3,8 @@ import 'package:voyager/voyager.dart';
 
 void main() {
   test('static values because codecov complains', () {
-    expect(RouterNG.cupertinoRoute, RouterNG.cupertinoRoute);
-    expect(RouterNG.materialRoute, RouterNG.materialRoute);
+    expect(Router.cupertinoRoute, Router.cupertinoRoute);
+    expect(Router.materialRoute, Router.materialRoute);
     expect(RedirectPlugin.KEY, RedirectPlugin.KEY);
     expect(WidgetPlugin.KEY, WidgetPlugin.KEY);
     expect(Voyager.KEY_TYPE, Voyager.KEY_TYPE);